move get_filter_info to utils.py
also add `expand_mask` to utils.py
